backend/requirements.txt
# ============================================================
# BACKEND DEPENDENCIES — pip install -r requirements.txt
# ============================================================

# --- Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- ML / AI ---
torch==2.4.1
torchvision==0.19.1
transformers==4.44.2
accelerate==0.34.2
sentencepiece==0.2.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
scipy==1.14.1
Pillow==10.4.0

# --- Transcription ---
openai-whisper==20231117
# OR use faster-whisper for speed:
# faster-whisper==1.0.3

# --- Database (dummy/dev) ---
tinydb==4.8.0

# --- Utilities ---
httpx==0.27.2
aiofiles==24.1.0
python-dotenv==1.0.1
loguru==0.7.2

# --- PDF Report Card ---
reportlab==4.4.0

# --- CORS ---
# (built into fastapi)
